Made the following modifications: 1. Converted ScalarTransportEquation to template form so that it can be used for both staggered and non-staggered scalar fields. 2. Added proper functionality to the Wasatch problem specification to accommodate for staggered scalar transport. 3. Converted ScalarRHS to template form so that it conforms to other expressions that we have (i.e. DiffusiveFlux …). 4. Fixed wasatch input files.

+++ CCA/Components/Wasatch/TimeStepper.cc
@@ -205,13 +205,13 @@ namespace Wasatch{
   {
     //__________________
     // loop over patches
-    for( int ip=0; ip<patches->size(); ++ip ){
+    for( size_t ip=0; ip<patches->size(); ++ip ){
 
       const Uintah::Patch* const patch = patches->get(ip);
 
       //____________________
       // loop over materials
-      for( int im=0; im<materials->size(); ++im ){
+      for( size_t im=0; im<materials->size(); ++im ){
         
         const int material = materials->get(im);
